requirements.txt
flask
flask-sqlalchemy
requests
bootstrap-flask
python-dotenv

